requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy==1.4.42
aiofiles>=0.8.0,<0.9.0
bcrypt>=3.2.0,<4.0.0
passlib>=1.7.4,<2.0.0
argon2_cffi>=20.1.0,<21.0.0
mysql-connector-python